Reorganize TaskNode.tsx
